refactor(tests): divide tests into small parts
